refactor(types): split up types in new `types/` folder
For better separation between pack and api types.
