+++ helpers.ts
@@ -1,5 +1,5 @@
 import * as coda from "@codahq/packs-sdk";
-import * as types from "./types";
+import * as types from "types/pack-types";
 
 export const ApiUrl = "https://photoslibrary.googleapis.com/v1";
