+++ params.ts
@@ -1,7 +1,7 @@
 import * as coda from "@codahq/packs-sdk";
 import * as helpers from "./helpers";
 import * as schemas from "./schemas";
-import * as types from "./types";
+import * as types from "types/pack-types";
 
 export const MediaDateRange = coda.makeParameter({
   type: coda.ParameterType.DateArray,
